Keep track of the iocage version our pool is compatible with.
This simplifies the handling of an edge case where iocage expects a
property that does not exist on datasets created by a different
version of iocage. Rather than translating unset values to "none"
in every get, check the pool version when iocage is invoked and set
the missing properties to "none" if the version does not match.

This commit also changes iocage to use a custom property on the pool
for activation, rather than hijacking the pool comment property.