Rewrite __hack88_mount. Bring back tag symlinks. Checks in set.
Rewrite __hack88_mount
----------------------
__hack88_mount was a scary name and was due for a makeover. To keep tag
symlinks pointed at the right place when hack88 is changed on a jail,
the links are now replaced whenever the mountpoint is checked for hack88.
The rename helps convey that the mountpoints are changed by the function.
The new name is __update_mountpoint and it wraps several helpers which
handle truncating or restoring the mountpoint and updating the tag symlink.
The new __update_mountpoint function is called when a jail is created,
cloned, or started. Changing the hack88 property on a running jail now
results in a fatal error, as it would change the mount point under a
running jail, which can't be a good idea.

Bring back tag symlinks
-----------------------
Previously tag symlinks had been disabled as they were not properly kept
up to date with the mountpoint of the jail, among other more mysterious
behaviors. With the rewrite of __hack88_mount, tag symlinks are now back.
Tag symlinks are now kept up to date when the mount point is changed.
When the tag property is changed by the user, the symlink is recreated.
When a jail is destroyed, the symlink is removed. Pretty simple.

Checks in set
-------------
Certain read-only ZFS properties are in ZFS_CONF and thus can be set by
the user. This was fine as they actually are read-only and zfs so would
provide an error message when they were attempted to be set. However we
also have properties that are read-write with zfs but should never be
changed by the user, specifically the mountpoint property. The global
defaults for these properties are already set to the string "readonly"
in lib/ioc-globals, so a simple check has been added in __set_jail_prop
to error out if the user tries to set one of these properties.