Rewrite list subcommand.
Overview of changes
-------------------
* Major rewrite of the `list` sub-command
* Long output, using -l instead of --long
* Getopts option parsing
* -H option to suppress headers
* -n option to add non-iocage jails to the listing, rather than by default
* Listing of releases and non-iocage jails moved to their own functions

Major rewrite of the `list` sub-command
---------------------------------------
The __list_jails function previously was deceptively simple. Unfortunately,
it was not flexible and extending the set of options it could handle was
not practical. To simplify the task, a getopts powered option parsing
routine has been implemented in lib/ioc-cmd. This routine passes a single
string in a fixed format to describe the output options requested by the
user. Pattern matching against this string of switches is performed using
`case` statements, allowing complex behavior to be codified in a concise
manner. Comments have been used extensively to clearly describe the behavior
of the various options and how they should interact. The output of the list
now pipes through column -ts^ in order to ensure a nicely formatted table.

Long output, using -l instead of --long
---------------------------------------
Introducing --long as the switch for long output was a complete break from
the usual short switches used everywhere else in the program. It also is not
compatible with getopts, so instead -l has been chosen for the flag. The
flag enables the display of additional information when listing iocage jails
and is a no-op for non-iocage jails and for releases. Upstream also rocked
the boat with the formatting of the long output. Here again a more customary
approach has been taken, opting for a header line and whitespace-separated
columns formatted by column -ts^ into a nice, aligned table.

Getopts option parsing
----------------------
Parsing options with getopts means compatible flags can be combined in any
arbitrary order and style. Eventually all the commands accepting multiple
options should use getopts for parsing.

-H option to suppress headers
-----------------------------
Similar to the `zfs` tools, -H can be passed as an option to this command in
order to suppress the header line.

-n option to add non-iocage jails to the listing, rather than by default
------------------------------------------------------------------------
Formerly the default behavior of the `list` sub-command was to display non-
iocage jails in a list after the listing of iocage jails. This has been made
optional. Listing non-iocage jails is now enabled by passing the -n option.

Listing of releases and non-iocage jails moved to their own functions
---------------------------------------------------------------------
Listing releases and listing non-iocage jails are in no way related to
the listing of iocage jails and have been promoted to their own separate
functions. It's possible listing releases may get its own sub-command in
the future.