Correct behavior of the -t flag for list. Doc fixes.
In reimplementing the __list_jails function, the behavior of the -t
flag was misread as behaving similarly to -b, in other words that
templates would normally be shown and the -t flag would only show
templates. The correct behavior is for templates to normally be
hidden, and the -t flag is used to show only templates. This change
adopts that correct behavior.

Documentation of the above behavior has been updated in comments and
in lib/ioc-help. A light style consistency pass has been made over
the whole help message in that file, as well.

The github url in the lib/ioc-help message under the BUGS section has
been updated to direct users to the freqlabs repo, as it is unlikely
bug in this version of iocage will apply to the upstream repo.