Allow read-only operations as a regular user. More.
Root checks
-----------

The opposite strategy from upstream is used in an attempt to have
reasonably simple logic while providing for comprehensive coverage
for checking that we have root when we need.

A helper __require_root has been created to be used as a fence in
lib/ioc-cmd before entering the parser for commands that require
root privileges. This puts the check in a highly visible place
where it will be hard to forget when adding new commands in the
future. The checks are also needed in the basic sanity checks on
the pool: __find_mypool, __check_filesystems, and
__check_defaults_ioc_version. When one of these checks requires
root privileges to activate a pool or create missing filesystems
or properties, a message explaining the requirement is displayed
before the check for root is performed. This ensures that running
a read-only command as a regular user while conditions cause one
of those checks to fail does not cause an unexplained error.

More
----

An __info helper in the style of __fatal_error but with the prefix
"INFO: " and without the exit has been added. This will be making
its way through the code to make informational messages consistent.

The df command has been given a getopts parser and a shiny new -H
flag for suppressing the header line.

The __print_disk_parseable function has been renamed to
__print_disk_parsable to fix the spelling error.

The __print_defaults function has been greatly simplified. There is
no reason to check the .defaults dataset as a fallback for the
global defaults. If the user wants to see the defaults in .defaults,
they must use `iocage get all default`. Global defaults set to none
have that value intentionally, and missing global defaults would be
a bug, so there is no reason to be looking on disk for any property.

The __show function has been relocated from lib/ioc-help to
lib/ioc-info.