New resolver property behavior.
The default value for the resolver property is now "copy" and the
behavior is the same as the previous default ("none"). The host's
/etc/resolv.conf is copied into the jail every time the jail is
started.

The behavior of "none" is now to do nothing with the jail's
resolv.conf. This was previously not possible, as iocage would
overwrite the file every time the jail starts.

The third behavior is unchanged. The resolver property may be set
to the contents of a resolv.conf file, with ';' substituted for
'\n'.