Template fstab, -F for clone, snapshot fixed.
The fstab file for a jail is now used as a template. All occurences
of %jailroot% will be replaced with the path of the jail's root
directory and the result is placed in /tmp/iocage/fstab.UUID, so
when the mountpoint of a jail changes due to hack88 the fstab still
works. It also makes fstabs more reusable, more readable, and easier
to write.

The documentation was lacking a description of the fstab file, so a
description has been added, including an explanation of the new
template string. This documentation is currently only in
lib/ioc-help.

The clone sub-command now has a getopts parser and will accept -F
as an option to keep the source fstab when cloning a jail, instead
of backing it up and creating an empty fstab in the new jail. This
has been documented in lib/ioc-help, as well.

The shapshot sub-command had a bug in the command processor when
a custom shapshot name was specified. It would cause the command
to fail because of an extra @ in the snapshot name. This has been
fixed so the extra @ is removed as intended.