Light consistency pass for getopts style.
Fixes a few places where the ? case was erroneously left unquoted.
Getopts sets a literal `?` as `_opt` for unmatched options. An
unquoted `?` is actualy a shell glob wildcard character.

Also fixed some places where `_opt` was used in the error message
for unmatched options. Again, `_opt` in these cases is set to a
literal `?`. The actual option character is set in `OPTARG`.