Add term color support detection to ioc-globals
Colors are nice and useful. Some useful strings have been created
as globals in lib/ioc-globals, along with a function to detect
color support in the terminal. If colors are supported, colored
versions of the strings are used. Otherwise, plain strings are
used.

The colored strings require echo to be given the -e flag. This may
deserve a wrapper eventually.