New clone -P option. Lots of bug fixes.
`__configure_jail` has been slightly tweaked to allow an additional
argument to be passed as the source dataset for cloning properties
from an existing jail rather than using the .defaults dataset. The
presence of this argument as a non-empty value alters the behavior
of certain properties which require different treatment depending
on whether the operation is a clone or a creation.

In combination with the update to `__configure__jail`, the switch
for keeping source properties in `__clone_jail` has been wired in
to use the new functionality of `__configure_jail`. Passing `-P` to
the clone command preserves the properties of the source jail, with
the exception of read-only properties, unique jail properties such
as host_hostuuid, and template which is assumed to always be "no"
for the clone. The doc string in `lib/ioc-help` has been updated
accordingly to describe this new flag.

Also in relation to `__configure_jail`, all calls to this function
have been appended with a condtional fatal error in case the
configuration step fails. This is necessary when the config is run
inside a subshell so that execution aborts properly.

Cloning a basejail copies several extra filesystems that need to
be cleaned up for the jail to start. This has been added to the
`__reclone_basejail` function by destroying not only old snapshots
of the release base, but also any leftover datasets in the way of
the base filesystems to be re-cloned.

Several string operations on snapshot names in the command module
have been corrected so that specifying a snapshot works now.

The show command will now abort when an unsupported property causes
`__get_jail_prop` to fail, instead of listing all the jail datasets
with blank values for the property.