Fix release updates, make messages consistent
A mistake in the path to freebsd-update.conf caused __fetch_release
to skip updating the release to the latest patch. This has been
fixed, so now iocage fetch will sucessfully patch the release.

Some additional helpers for message logging to stderr have been
implemented and made use of in various places. This helps the
messages keep a more consistent style, and places informative
info on stderr instead of stdout.