Refactor __create_jail, allow comments in pkglist.
`__create_jail` needed to be restructured because the program flow
was beginning to get complicated. Each of the cases in the jail type
flag switch statement deserves its own function. These blocks are
fully self-contained and putting them outside of `__create_jail`
makes both areas of the code simpler to read and comprehend. The new
functions broken out of the case switch are:
- `__create_jail_clone_datasets`
- `__create_jail_empty_datasets`
- `__create_jail_basejail_datasets`
- `__create_jail_standard_datasets`

Also broken out of `__create_jail` is a new helper function,
`__require_release_fetched` for... requiring that a release has been
fetched. Related to this function, a fetched release is no longer
required to create empty jails.

The `__pkg_install` function got a makeover so that the correct
mountpoint of the jail is used to form the chroot path for pkg,
grep is used to filter out lines commented with a `#` from the
pkglist file, and all the package names from the pkglist file are
passed to pkg in once invocation, instead of looping over the list
of packages one by one.

A new function named `__jail_fstab` has been added to
`lib/ioc-configure`. This function spits out a few lines of comments
to explain the fstab file and the %jailroot% keyword. This block of
text is now written to the fstab file instead of leaving the file
blank when creating or cloning a jail.

Jail lists are now sorted by tag, simply to make lists look nicer.

Cloning a jail, iocage will now check for the existence of rc.conf
in the jail before attempting to copy it to the new jail and update
occurences of the old UUID to the new UUID. This prevents an error
when cloning custom jails that do not have an rc.conf.