Use `host_hostname` and eliminate `hostname`.
The `hostname` property broke convention for the naming of jail
properties, is not documented for iocage, and was not exposed at
creation time to the user. `host_hostname` is the conventional,
documented name for this property. We expose this property on the
command line when creating or cloning jails. Using `host_hostname`
would have no effect before this change. Now the property behaves
as documented, and `hostname` is no longer used for anything, so
has been removed.

This commit resolves GitHub issue #9.

Some comments in `__configure_jail` have also been reworded for
clarity.