Fix broken version check. Add paranoid checks.
A copy-paste error in `__check_dataset_ioc_version` causing an unset
variable to be passed to `__update_dataset_ioc_properties` as the
dataset has been fixed. Unused locals in the version check related
functions have been pruned, and a loop variable has been renamed
from `_jail` to `_dataset` in `__check_all_ioc_version`.

`__check_all_ioc_version has been added to a few places where `ALL`
is used as a keyword, though this check is not really necessary at
this time as these commands do not currently make use of any new
properties.

`__check_dataset` has been updated to only check the dataset ioc
version if it is not the .defaults dataset. The defaults have been
checked already by the time this function is called, and it is not
sane to potentially call `__update_dataset_ioc_properties` on the
.defaults dataset, as this function looks to .defaults for the
values it sets.

In `lib/ioc-common`, the helper function `__check_dataset_all` has
been removed, as it is not used, and it is safer to instead call
`__check_dataset` inside the loop whenever `__check_dataset_all`
might have been used before a loop.