Block "ALL" and "default" tags. Improve messages.
"ALL" and "default" are not allowed to be the tag. This is now
enforced in `__export_props` and `__set_jail_prop`.

Messages when starting/stopping/restarting jails are now less
verbose. Errors during these operations now  mention the UUID of
the jail.

Properties set on the command line now undergo a more stringent
test before being exported in `__export_props`.