Support snapremove ALL@*, improve @ALL.
iocage snapremove ALL@ALL
                  ALL@snapshot[%snapshot][,snapshot[%snapshot]]...
                  uuid@ALL
                  uuid@snapshot[%snapshot][,snapshot[%snapshot]]...
                  tag@ALL
                  tag@snapshot[%snapshot][,snapshot[%snapshot]]...

Ranges and lists of snapshot should just work as they are part of zfs. This still
needs to be documented.

snapremove (ALL|uuid|tag)@ALL avoids trying to remove clones so that zfs does not
spam the user with "not removing because clones" messages. This is not simple
to check in cases other than @ALL without removing the ability to use ranges and
lists of snapshots.

There should be no worries about destroying clones or basejails. Clones block
snapshot deletion. Basejails are just short lived clones, and the origin
snapshots aren't even in the same place as the jails.