Warn that the origin snapshot may be destroyed.
When destroying a clone, it is possible that the origin snapshot
may be destroyed as well, if no further clones are left of that
snapshot. This is potentially undesirable. Warning the user is the
least we can do until a more correct behavior is implemented.

Also remove a bunch of indentation in messages to give ourselves
more room. UUIDs take up a lot of space in a line, as it turns out.

The message format still needs unity and polish, and some things
got shuffled around between stderr and stdout while entertaining an
idea.

The messages when applying rctl limits to a jail now mesh better
with the rest of the messages during jail startup.

A bug whereby a failed upgrade of a basejail would not restore the
previous release has been spotted and fixed.