Change default for resolver to be non-destructive.
It can be very unexpected that resolv.conf is clobbered each time a
jail is started. Setting the default to a non-destructive behavior
makes more sense than copying the host's resolvers by default,
which often makes no sense at all (particularly with the adoption
of local unbound). The user can set their own default to "copy" if
the old behavior is desired.

Some additional detail and fixes in ioc-help accompany this change.