Cleanup in legacy start code, fix *_addr params.
This was a bit of putting the horse back in front of the cart.

The previous doc update turned out to be unsupported, until now.
ip4_addr and ip6_addr are now fully tolerant of spaces. IFS is used
in a controlled fashion to split the address lists properly when
looping.

The legacy start code also had to be reworked to properly quote the
parameters. In doing so, it was finally decided to pull out the ipv6
global. It wasn't tunable, and removing it simplifies the legacy
start logic.

Unfortunately there seems to be a bug in jail(8) that breaks extra
params when a netmask is specified using the / notation.