Fix error in term color test, ip*_addr spaces.
When the terminal doesn't support colors (during rc boot for example)
the test as written would result in a shell syntax error and cause a
cryptic error message to be shown. No other ill effects manifested.
The test has been rewritten so that when the first test fails the
second test (which would be invalid) is skipped.

jail(8) handling of spaces in ip[46].addr is very immature. A space
following a comma in a list is not taken gracefully. To be more
accomodating, a string substitution replacing ', ' with ',' has
been introduced in the legacy start function.

__stop_legacy_networking has been given a minor makeover, opting to
use the handy __sub_default_iface function to handle the string
substitution of replacing DEFAULT with the actual interface. Some
extra stripping of optional parameters from the addresses before
removing aliases has also been added, along with helpful comments.