Remove record, package, import, export, init-host.
These features were all implemented in ways that did not take any
advantage of ZFS's capabilities. Record, for instance, is basically
an incremental snapshot, and zfs diff can even add a nice human
readable representation of changes. Package seemed redundant or out
of place. Import and export missed an obvious win by not using zfs
send/recv. Init-host seemed pointless.

The plan is to eventually implement iocage send/recv, ideally in a
way that retains the ability to pipe around a data stream. Record/
package are not a priority currently, but may be reevaluated in the
future.

Also docs have been figured out! Copying the text from ioc-help
into iocage.8.txt and then the key observation that txt2man was
used to generate iocage.8 means that all the docs are now up to
date with the changes in ioc-help! This is a huge help and should
encourage more thorough documentation in the future.